requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
ultralytics==8.0.196
